WIP Refactoring provider configuration with directory and database sources
