+++ src-tauri/src/directory/commands.rs
@@ -1,12 +1,26 @@
 use super::models::{Directory, File};
 use crate::content::containers::HeterogeneousContentArray;
-use crate::directory::helpers::get_file_contents;
 use crate::error::DwataError;
 use crate::relational_database::crud::CRUD;
+use crate::workspace::configuration::{Configuration, ConfigurationData};
 use crate::workspace::DwataDb;
 use log::error;
 use std::path::PathBuf;
 use tauri::State;
+
+#[tauri::command]
+pub async fn create_folder_source(
+    path: &str,
+    label: Option<&str>,
+    include_patterns: Vec<&str>,
+    exclude_patterns: Vec<&str>,
+    db: State<'_, DwataDb>,
+) -> Result<String, DwataError> {
+    match *(db.lock().await) {
+        Some(ref mut db_connection) => Directory::create_configuration(data, db_connection).await,
+        None => Err(DwataError::CouldNotConnectToDatabase),
+    }
+}
 
 #[tauri::command]
 pub(crate) async fn fetch_files_in_directory(
@@ -46,7 +60,7 @@ pub(crate) async fn fetch_file_contents(
                     // Find the FolderSource matching the given folder_id
                     let full_path: PathBuf = directory.path.join(relative_file_path);
                     if full_path.exists() {
-                        Ok(get_file_contents(&full_path))
+                        Ok(Directory::get_file_contents(&full_path))
                     } else {
                         Err(DwataError::CouldNotOpenFolder)
                     }
