+++ src-tauri/src/directory/mod.rs
@@ -1,5 +1,5 @@
 pub mod commands;
-// pub mod configuration;
+pub mod configuration;
 pub mod crud;
-pub mod helpers;
+pub mod file_contents;
 pub mod models;
